# api-specs/policy-service.openapi.yaml
openapi: 3.1.0
info:
  title: Policy Service API
  description: Policy management and enforcement service for Rust Security Platform
  version: 1.0.0
  contact:
    name: Security Team
    email: security@rust-security.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.rust-security.example.com/policy
    description: Production server
  - url: https://staging-api.rust-security.example.com/policy
    description: Staging server
  - url: http://localhost:8002
    description: Local development
security:
  - bearerAuth: []
  - apiKey: []
...
paths:
...
  /v1/authorize:
    post:
      tags:
        - Evaluation
      summary: Authorize a single request using Cedar policies
      description: |
        Authorize a request. The `action` field expects a compact identifier describing the action to evaluate. Preferred format is "Domain::verb" (for example, "Document::read"). The service will attempt to parse the Action into a Cedar Action entity; parse failures return HTTP 400 with an error_type of "authorization_error".
      operationId: authorize
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorizeRequest'
            examples:
              basic:
                summary: Basic authorize example
                value:
                  request_id: "req-123"
                  principal:
                    type: "User"
                    id: "user:alice"
                  action: "Document::read"
                  resource: "doc:42"
                  context:
                    client_id: "service-x"
      responses:
        '200':
          description: Authorization decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorizeResponse'
        '400':
          description: Invalid request parameters (e.g., action parse/validation failure)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
...
  schemas:
...
    AuthorizeRequest:
      type: object
      required:
        - request_id
        - principal
        - action
        - resource
      properties:
        request_id:
          type: string
          description: Unique request identifier for tracing
        principal:
          type: object
          description: Cedar entity describing the principal (type & id)
          additionalProperties: true
        action:
          type: string
          description: |
            Action identifier. Preferred format is "Domain::verb" (e.g., "Document::read"). The service will accept simple values like "read" but recommends the namespaced form for clarity.
          example: "Document::read"
        resource:
          type: object
          description: Cedar entity or identifier for the resource being accessed
          additionalProperties: true
        context:
          type: object
          description: Additional contextual attributes (e.g., client_id, ip address)
          additionalProperties: true
    AuthorizeResponse:
      type: object
      properties:
        decision:
          type: string
          description: Decision string ("Allow" or "Deny")
...
    ResponseMeta:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        api_version:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
            field_errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...